change notebooks to scripts
